unify fluid blocking queue
